requirements.txt
requests          # To make HTTP requests to Gemini API
pandas            # For any data manipulation needs (optional)
flask             # If you want to create a web app or API frontend
python-dotenv     # To handle environment variables loading (.env files) with getenv
google genai